Move PhotoWrapper in another file. #55
